Modified VTheta.m
I had concerns that squaring the product of the X_i’s would include
terms that were cross terms between quadratures (which should not be
correlated).  In ordinary vector-matrix operations, this would be fine,
but we are actually manipulating random noise variables (the cross
terms really matter here). By element wise squaring the THETA_i
matrixes first and then matrix multiplying by the input (uncorrelated)
noise terms (typically vacuum) this ensures that all cross terms are
dropped.

This means that the model can only deal with uncorrelated inputs, but
that should be ok for now.  Would need to do some serious thinking if
we wanted to inject squeezing into a squeezer.